IQSD-1931: Фиксы message.vue
